refactor(frontend): reorganize feature modules
