+++ crates/fricon-ui/frontend/src/features/chart-viewer/components/ChartWrapper.tsx
