+++ crates/fricon-ui/frontend/src/features/chart-viewer/hooks/cascadeReducer.ts
