+++ crates/fricon-ui/frontend/src/features/chart-viewer/hooks/useDatasetWriteStatusQuery.ts
