+++ crates/fricon-ui/frontend/src/features/chart-viewer/hooks/useFilterTableDataQuery.ts
