+++ crates/fricon-ui/frontend/src/features/chart-viewer/index.ts
@@ -1,0 +1,1 @@
+export { ChartViewer } from "./components/ChartViewer";
