+++ crates/fricon-ui/frontend/src/features/chart-viewer/lib/chartViewerLogic.ts
