+++ crates/fricon-ui/frontend/src/features/data-explorer/index.ts
@@ -1,0 +1,1 @@
+export { DataViewer } from "./components/DataViewer";
