+++ crates/fricon-ui/frontend/src/features/dataset-detail/index.ts
@@ -1,0 +1,2 @@
+export { DatasetDetailPage } from "./components/DatasetDetailPage";
+export { useDatasetDetailQuery } from "./queries/useDatasetDetailQuery";
