+++ crates/fricon-ui/frontend/src/features/dataset-detail/queries/useDatasetDetailQuery.ts
