+++ crates/fricon-ui/frontend/src/features/dataset-table/hooks/useDatasetTableData.test.ts
@@ -2,7 +2,7 @@ import { createElement, type ReactNode } from "react";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { act, renderHook, waitFor } from "@testing-library/react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { useDatasetTableData } from "@/components/use-dataset-table-data";
+import { useDatasetTableData } from "@/features/dataset-table/hooks/useDatasetTableData";
 import type { DatasetInfo, ListDatasetsOptions } from "@/lib/backend";
 
 type ListDatasetsFn = (options?: ListDatasetsOptions) => Promise<DatasetInfo[]>;
