+++ crates/fricon-ui/frontend/src/features/dataset-table/index.ts
@@ -1,0 +1,1 @@
+export { DatasetTable } from "./components/DatasetTable";
