+++ crates/fricon-ui/frontend/src/features/workspace/index.ts
@@ -1,0 +1,1 @@
+export { useWorkspaceInfoQuery } from "./queries/useWorkspaceInfoQuery";
