+++ crates/fricon-ui/frontend/src/features/workspace/queries/useWorkspaceInfoQuery.ts
